Add location to Substring in AST
Summary:
The mixed option for strings may contain expressions embedded in format strings. Currently, our string AST keeps track of:

1. location of the entire string, quotations and prefix included (we do want to keep this location around as the location of the string expression, because the quotations are indeed part of the string itself; the unquoted expression would mean something different)
2. the contents of the string without prefix and quotation (we also want to keep the content quotationless, because we'd like to be able to interpret string objects as a string annotation, etc., and keeping quotations or prefixes around would require us to parse out the true content every time we looked at it)

This combination of things we store mean that it's impossible for us to reconstruct the location of nested expressions. We *did* have logic that purported to do this, but it was spitting out nonsense, really (see existing test cases that got changed in D15926415).

We can't just hard code an offset because we don't know whether a string was constructed with single quotations, double quotations, or had a 1- or 2-character prefix, etc.

The best way to fix this, imo, is to add a location to the substrings we collect, since there may be expressions embedded inside and we do care about these locations.

Reviewed By: dkgi

Differential Revision: D15926080

fbshipit-source-id: a9741b9a7913b150f8f7e603d3767bd7e770faf2