Fix locations in of expressions in format strings
Summary:
Our preprocessing step for handling format strings had a lot of bogus logic for determining the locations of the expressions it parsed out.

In every case, these locations were wrong - they did not take into account the prefixing or quotations that the string was defined around, so always pointed to some piece of the string that was too early.

Keep track of more nuanced location information in the lexer/parser so we have enough info accurately reconstruct the locations of nested expressions.

Newlines are still broken because they are not taken into account in the existing logic, but I added a test that demonstrates the breakage and will fix this next.

Reviewed By: dkgi

Differential Revision: D15926415

fbshipit-source-id: 93aa9b72b292f7a2fac455c509d7c7f25528e168