Introducing ParameterSpecification components
Summary:
As discussed in their introduction in D15339143, `ParameterSpecification` variables couldn't be safely called.  This diff introduces a mechanism to soundly call them, or to synthesize one.

The key to the approach is the idea of a nested function definition as being a "literal" of a kind.
# Enclosing Scope
From the enclosing scope's perspective, the problem is a matter of trying to make a new instance of the type `Callable[TParams, Something]`.  Previously, there was no way to ever make one of these that wasn't from your parameters, or originated as from a cast.
With this diff I introduce a syntax for a "literal" of the type `Callable[TParams, Something]`.
That syntax is (exactly, mod names for args and kwargs) the following:
```
def outer(..., x: Callable[TParams, int], ...) -> ...:
  def inner( *args: PositionalArgumentsOf[TParams], **kwargs: KeywordArgumentsOf[TParams]) -> Something: ...
```
this syntax will add `inner` -> `Callable[TParams, int]` to the local resolution.
Note that it is not adding it as `Callable[[Variable(PositionalArgumentsOf[TParams]), Keywords(KeywordArgumentsOf[TParams])`.
These type expressions are only valid in exactly this form in exactly this location.  This is because there is no reasonable interpretation of the meaning of one without the other, or a signature with anything else included.
This is also only accepted inside a function that already binds that type variable.
This can be understood as a consequence of the idea that you can only annotate locals with bound type variables (which is true).

# Inner Scope
From the inner scope's perspective, the problem was being able to safely call the `Callable[TParams, Something]`.
The solution now is that for functions with the signature `( *args: PositionalArgumentsOf[TParams], **kwargs: PositionalArgumentsOf[TParams])`  (exactly, mod names for args and kwargs) , we initialize their resolutions with the types `args` -> `PositionalArgumentsOf[TParams]` and `kwargs` -> `PositionalArgumentsOf[TParams]`.  These are in fact types, as we want to flow this identity through assignments, etc.  I had to add a new variant since this kind of type has a different contained data than any other type, deserving a distinct variant.
These are entirely opaque types, which ought not to be used for anything other than being ferried over to a call to the function.
This means that there is now exactly one way to call a `Callable[TParams, Something]`, namely `*(PositionalArgumentsOf[TParams]), **(KeywordArgumentsOf[TParams])`.

# Architechture

This is a pretty complicated feature which had the potential to introduce a lot of footguns.  I tried to seal away all of the complexity inside of Type.Variable.Variadic.Parameters.  The concept of the interface looks like this:

{F162791194}

`create_overload` stops at `parse_instance_annotation` because the public interface of the function only needs the variable.
`TypeCheck.initial` needs to both parse the declaration, then `decompose` the variable into the two types in order to allow them to flow towards a function invocation.
At every function invocation `AnnotatedSignature.select` attempts to `combine` the types back into a variable, which then can be compared against the expected one.
At all of these stages, externally to the `Type` module, the internals and identity of the components are completely opaque.

Reviewed By: shannonzhu

Differential Revision: D15914752

fbshipit-source-id: 878b76765c772851b331c44a0b448cc3eeb4cf7a