Make nested functions no longer parts of global scope
Summary:
I think as a convenience, nested functions were being registered into the global namespace just like any other function.  This was problematic for a variety of reasons:
* These nested functions are not actually globally accessible in the runtime, so accesses on them externally ought to fail, just like any other local variable
* Nested functions live inside of the type variable scope of their parent function, and so type variables that originate from those enclosing scopes ought to be understood as bound variables, not free ones (this was the inciting problem that triggered this work)
* registering them as globals means we allowed you to use them before declaring them

I fixed these problems by changing preprocessing to qualify them as locals rather than globals, changed environment building to skip them during registration, and changed type checking to add them to the resolution at their declaration point, with the correct type variables.

Doing this exposed some broader issues with our approach to nested functions.  The main issue at play here is that from the perspective of a nested function, variables coming from the enclosing scope are akin to global variables.  Our approach right now is to schedule analysis on a nested function at the point of declaration, with the resolution as it is in that moment.  However, since those outer variables are considered local by the enclosing function, we allow their types to change in between the declaration and invocation of the function, which is potentially problematic
```
def outer() -> None:
  x = 7
  def inner() -> None:
    expects_int(x) # type checked when x is still an integer
  x = "A"
  inner()
```

Reviewed By: shannonzhu

Differential Revision: D15909160

fbshipit-source-id: 092777088936b300cfeedd1b0eac947c98840286