Store type check resolutions for open documents in the server
Summary:
Right now, we don't store type check resolutions for the Pyre server at all, which hinders our ability to serve accurate hover annotations. Make it so that the Pyre server's open documents always store the type check resolution.

An optional argument `open_documents` is added to `Check.analyze_sources`. This argument passes information to `Check.run` to override storing type check resolution strategy if the file to be type checked is in the `opened_documents` set. The argument is default to false to maintain the old behavior of not storing resolution by default. An extra `unit` argument is added to the function to enable the usage of the default argument.

Reviewed By: grievejia

Differential Revision: D15931916

fbshipit-source-id: 1550bdf03fea578f62ad0eea5bbec62e9bb23af3