Add a `ModuleTracker` service to keep track of module-to-source mapping
Summary:
The Problems
==========
One module name may correspond to multiple potential source files. Various reasons contribute to this. One example is the simultaneous existence of src/a.py and src/a/__init__.py. Another example is the simultaneous existence of ext/a.py and src/a.py, where ext/ is in an external search path and src is in the project source directory. This is a problem that cannot be solved by simply banning directory overlaps since (1) Source file overriding behavior is sometimes intentional. E.g. we want to override an undesirable typeshed stub with our own stub. (2) Many open-source users have a setup where `virtualenv` or something similar lives under the project root. We don't want to create obstacles for those users when they try to adopt pyre.

Our existing solution to this problem is to assign an order to all files that correspond to the same module, pick the first file according to that order, and discard all info about the result of the files. This solution has 2 drawbacks:
- The order which we assigned to those files relies on an unreliable filesystem directory traversal order. D15401856 partially addressed this for typeshed, but we still need to enforce a stronger order for other cases.
- Incremental changeset computation is not well-supported. Currently if we remove a file, incremental analysis always treats the module as "deleted", without considering the possibility that there might exist other files in the search path that also correspond to the module. This lead to server inconsistencies.

NOTE: Besides, it would also be nice if we have a proper abstraction layer that separates filesystem plumbing and the module structure that the type checker sees.

The Solution
==========

This diff does two things:
- Assign a more explicit order to files that correspond to the same module.
- Instead of recording only the "winning" file for each module, we keep track of all "losing" files as well.

Both of these were done within the newly introduced `moduleTracker.ml`. The tracker is simply a map from module names (i.e. `Reference.t`) to a list of files it corresponds to. The file lists are sorted in an order which we prioritize them. To make it more pleasant to program with, information that is required to make the prioritization decisions are wrapped in a structure named `SourceFile`. `SourceFile.same_module_compare` is the function where the ordering is implemented, which should be fairly self-explanatory. `SourceFile.create` does the job of walking over the search path and try to locate the file and assign priorities.

`ModuleTracker.create` tries to populate the tracker by (1) looking for all files that pyre cares about under the search paths, (2) making sure that they fall into the right module bucket, and (3) for files under the same bucket, making sure that they are sorted according to `SourceFile.same_module_compare`. This function is intended to replace `Service.Parser.find_stubs_and_sources`, but I haven't hook it up to the production code because additional cleanup work on the parser side is needed (e.g. setting up a global tracker, and connecting `ProjectSpecifiPreprocessing` to it).

Reviewed By: dkgi

Differential Revision: D15424426

fbshipit-source-id: e5a9278e684f3d6fd982f53148e7ea45ef41fda5