Remove Location.any default parameter values
Summary:
I want to get rid of hidden `Location.any`s in all non-testing functions.

This already doesn't happen in most places, but imo it's a little scary to be able to call `Parameter.create` or `Reference.expression` and not need to think about the location or make it obvious to someone reviewing your change. An explicit Any will at least remind you to consider what the location should be.

Not sure if all of these callsites are fine, but at least make them explicit for now and I'll add integration testing for locations to check.

Reviewed By: dkgi

Differential Revision: D15932392

fbshipit-source-id: 634cd89bf58151c8bfd42eaaebcca5ee772a8187