Annotate iggraphql functions to start to get some coverage
Summary:
This diff adds support for generating taint models for graphql resolver functions. Not that the functions are always formatted with three args: `(obj, info, **args)`. `obj` is the root object of the query, and isn't user input. `info` is of the `ResolveInfo` type and will be addressed in D15764060. `args` is the user supplied input and is the only thing that we need to taint.

Note that this diff also modifies the existing parsing code in `generate_taint_models.py`. One major limitation of the previous code was that `ast.NodeVisitor` doesn't seem to recurse down through the graph. I added an entire tree walk so that the various cases of nested `GraphQLField`s (see my tests) wouldn't be missed.

Reviewed By: dkgi

Differential Revision: D15763907

fbshipit-source-id: 8bf9eacb26d3fd603dac9ec216b5f6194c708403