Merge multiple incremental update to the same module
Summary:
The goal of this diff and the previous one is to implement `ModuleTracker.update`, where we take a list of paths watchman sends us and produce a list of `IncrementalUpdate.t`, telling us which module has changed. More specifically, there are three different ways a module can change, and we need to be able to distinguish them:
- A new module gets created, or the underlying file of an existing module has potentially changed (`IncrementalUpdate.New`)
- An existing module gets removed (`IncrementalUpdate.Delete`)

The issue we are trying to address in this diff is to make sure the return of `ModuleTracker.update` only contains one update per module: if the filesystem events made multiple updates to the same module, we want to "merge" these updates together (i.e. finding a single update that is equivalent to the original multiple updates conducted in sequence). As a side bonus, such a merge procedure canonicalizes the results and make it insensitive to the order in which the filesystem events take place.

The merging algorithm we use here is best to think of as a state machine with 4 states: EMPTY, NEW, DELETE, and ERROR. EMPTY means nothing has happened to the module. NEW and DELETE corresponds to their counterpart in `IncrementalUpdate`. ERORR is the error state that should never have happened. Transition happens when we detect a new `IncrementalUpdate` for the said module:
- EMPTY goes to NEW or DELETE depending on what the `IncrementalUpdate` is.
- DELETE goes to NEW when seeing a `New` update (deleting the module then adding it back is essentially the same as changing the original module)
- NEW goes to DELETE when seeing a `Delete` update (this happens when we repeatedly delete the shadowing file for the module until every file gets deleted)
- NEW goes to NEW when seeing a `New` update (the second change will override the first)
- All other transitions go to ERROR (they should never happen in practice)

Reviewed By: sinancepel

Differential Revision: D15958470

fbshipit-source-id: f7eeaed716dbcfb790e85c9729bfe97de50ff7d6