Compute incremental update from filesystem event in ModuleTracker
Summary:
The goal of this diff and the next one is to implement `ModuleTracker.update`, where we take a list of paths watchman sends us and produce a list of `IncrementalUpdate.t`, telling us which module has changed. More specifically, there are 2 different ways a module can change, and we need to be able to distinguish them:
- A new module gets created, or the underlying file of an existing module has potentially changed (`IncrementalUpdate.New`)
- An existing module gets removed (`IncrementalUpdate.Delete`)

The first issue here is that watchman only tells us *which* files has changed. It doesn't tell us *how* the files are changed. But since we get access to the filesystem ourselves, we can simply query the existence of the changed files: if a file no longer exist, it must have been removed; if it still exists, it must be either added as a new file or updated as an existing file. Adding vs. updating a file do not make much of a difference to us, but it's important to distinguish between updating vs. removing.

The next issue is how to translate changes of files to changes of modules. If there's a 1-1 mapping between python modules and python source files, the problem would be easy: if a file gets updated and we didn't track it before, it must be a new module; if a file gets updated and we tracked it, it must be a changed existing module; if a file gets removed, it must be a removed module.

However, in Pyre the mapping between modules and files is not 1-1: there could be potentially many source files corresponding to a module, and one source file may "shadow" another source file if they get mapped to the same module. In D15424426 we attempted to solve this problem under the scenario of end-to-end full check. This diff can be seen as an attempt to solve this problem under the scenario of incremental check.

If we take shadowing into account, the basic algorithm described above needs to be adjusted accordingly: we need to consider the case when a file gets updated/removed but it gets shadowed by another file, and the case when a file shadowing other files gets updated/removed. In both cases, file changes do not translate directly to module changes.  See `ModuleTracker.process_filesystem_event` for details.

Reviewed By: sinancepel

Differential Revision: D15953191

fbshipit-source-id: 3a9715fb353dab698cd09f235a3b00e8dbea271c