Fix resolution of annotation literals
Summary:
`reveal_type(typing.List[int])` used to result in: `Revealed type [-1]: Revealed type for typing.List.__getitem__(int) is `typing.Any`.`

This is because many of these types are defined in `typing.pyi` as:

```
class TypeAlias:
    # Class for defining generic aliases for library types.
    def __init__(self, target_type: type) -> None: ...
    def __getitem__(self, typeargs: Any) -> Any: ...

Union = TypeAlias(object)
Optional = TypeAlias(object)
List = TypeAlias(object)
Dict = TypeAlias(object)
DefaultDict = TypeAlias(object)
Set = TypeAlias(object)
FrozenSet = TypeAlias(object)
Counter = TypeAlias(object)
Deque = TypeAlias(object)
ChainMap = TypeAlias(object)
```

When we try to resolve the expression inside the reveal type, we wind up with a call to `typing.TypeAlias.__getitem__` which has a return of `typing.Any`.

Not sure how we handled this before or if we did, but it seems to make sense that in the case resolving the expression inside reveal_type doesn't work out, we should try to parse it as a type itself.

Reviewed By: dkgi

Differential Revision: D15979922

fbshipit-source-id: ed870b9bd901bd5e8527421b14da1a26ae5275e8