Store annotation in forward resolved object
Summary:
Sometimes we care about retrieving the full annotation of an expression, not just the type it would resolve to if executed or passed to a function or composed in an expression at that moment. (Immediate use-case: reveal_type and lookups - existing use case is for the target annotation of an assign)

`resolve` gives a type, which is what we want in almost every case. In fact, the idea of an 'Annotation' only exists for names/identifiers, and not for any other kind of expression that doesn't carry state around.

We could either refactor all cases in forward expression, forward statement, and all users of resolve to return or expect an Annotation.t, and start wrapping/unwrapping types in Mutable Annotations as needed. This is ugly and adds a lot of overhead; let's just keep track of this extra bit of information that is sometimes useful by adding it to the resolved state.

Note: thought about adding another function to resolution by simply redefining logic that will expose annotations rather than types: D16007502

Reviewed By: grievejia

Differential Revision: D16008804

fbshipit-source-id: 9767554255ec2bca454d3ebbc7ad87913bac9b78