don't follow symlinks for `qualifier`
Summary:
For the `_qualifier` function, resolving symlinks unconditionally sometimes leads to issues. Consider:

```
/unrelated/a.py
/root/b.py
/root/subdirectory/a.py (symlink to /unrelated/a.py)
```

In this case, the qualifier we want for `a` is `subdirectory.a`, not `...unrelated.a`. Avoiding the symlink resolution here makes us handle this case properly.

The GraphQL model generation logic relied on this behavior, so normalize the path when generating those models in the caller's end.

Reviewed By: dkgi

Differential Revision: D16002096

fbshipit-source-id: 38ebf9ce291fe9f437cddd8023c75fcc4bad23c3