Limiting the typeorder to contain only Top, Bottom, Any and Primitives
Summary:
I want to get the typeorder converted over to only contain primitives.  The first step to this is to formalize with the type system what's already the case, which is that the type order contains `Top`, `Bottom`, `Any` and all of the primitives.
Next step is removing `Any`, which will require some work to make inheriting from something from a placeholder stub work as expected.
Then we should be able to remove bottom and top, as they should be implied to be connected to all nodes, but need not be actually connected to any of them.

Reviewed By: dkgi

Differential Revision: D16020128

fbshipit-source-id: 343c17907c9b68ad69767ba288825c79786b5820