Fix same module comparison order for stubs
Summary:
In D15929873 I changed the meaning of `SourceFile.is_external` from "is the source root of this file `local_root`" to "do we want to type-check this file", but I did not update the logic of `SourceFile.same_module_compare`, which depends on `is_external`. This violates our principle of separation: the decision of whether to type check the file or not should not interfere with the source file discovery algorithm.

This diff update the comparison logic to ignore `is_external`. I also reshuffled the comparison priority such that search path order takes precedence over whether the module comes from `__init__.py`/`__init__.pyi` or not: the philosophy here is that we should stick to the search path order more often to minimize user surprises. I am tempted to even prioritize the search path order over whether the module comes from `.pyi` or `.py`, but apparently that's going to break our current distillery settings so let's avoid getting too radical on this issue.

Reviewed By: dkgi

Differential Revision: D16002247

fbshipit-source-id: 14ff85d1c69e793d8a5b5afa188853bce2c11893