Removing Top and Bottom from the type order
Summary:
As described in D16020128, it is redundant having Top and Bottom explicitly included in the type order, so this diff removes them.
I also removed some unneeded flexibility in the poorly named functions that previously were only ever used in prod to deal with connections to `object`.
In theory we need not put `object` in the lattice either, though doing that will take a lot more work, and isn't within the scope of the current effort.

Reviewed By: grievejia

Differential Revision: D16035541

fbshipit-source-id: b0afa6696802030051af68f82d4d8c4fe73e6b70