Add resolve to annotation functionality to resolution
Summary:
Adding a lookup for the full annotation of an expression - if it's a name or identifier (aka, a value store such as a local or global variable or an attribute access), this might contain both a global annotation and a local annotation. Otherwise, all other expressions will evaluate to a mutable type.

See description of diff below for more context, and diff above for a use case (this became more urgent than just the reveal_type use case because we were exposing some confusing types at lookup query)

Reviewed By: dkgi

Differential Revision: D16015024

fbshipit-source-id: 2b039da7abecdbe02c90bc50cce696140d551520