Reveal full annotation in reveal_type
Summary:
We had reports that our reveal type was confusing in cases where we have a global annotation for an identifier but also a local version.

See: https://fb.workplace.com/groups/1372630086160153/permalink/2308036125952873/

Turns out we run into the same problem on type lookups for query and libCST integration, so this needs to be fixed soon anyway.

Reviewed By: sinancepel

Differential Revision: D16015271

fbshipit-source-id: 1b3e7f5c5798ee1b25632cb1da0b76ba9b8e545b