Use original type in globals for lookup where known
Summary:
When we process `x = 1`, we set the type of `x` to `int` but the local type of `x` to `Literal[1]`, meaning you won't get an error if you try to reassign `x` to another int, but you it is also acceptable to pass `x` in as an argument where a literal 1 is expected.

It gets confusing for reveal type to reveal the local inferred type rather than the annotated or general type of `x`, and same with type lookups.

Reviewed By: grievejia

Differential Revision: D16034017

fbshipit-source-id: bacf0b1740a754befb8dfe1a33f8df3cf4757de7