Process buck builder log output in Pyre client.
Summary:
Previously the Java logging library will interfere with Python logging in Pyre client that sometimes results in a Java log and Python log being merged into a single line.
This commit fixes the problem by processing the stderr of the Java process and use the Python logger to re-log to ensure consistency.

Reviewed By: shannonzhu

Differential Revision: D16043656

fbshipit-source-id: cea075436303facf6e138a77e17115638996bc81