Fix implicit submodule invalidation
Summary:
We weren't doing the submodule invalidation properly before: we a module needs to be purged in the environment, none of its implicit submodules were touched at all. This is a potential source of server inconsistency: if module `a.b.c` gets removed, and `a.b.c` is the only package under `a.b`, then we'll still pretend that `a.b` exists in the incremental mode while in full check mode we won't register `a.b` at all.

To address this issue, we explicitly keep track of the submodules and invalidate them in `Handler.purge`. Note that it is not enough to just record whether the implicit submodule exists or not --- implicit submodules can be shared among multiple non-implicit modules. Only when all of its non-implicit modules go away can we safely invalidate the shared implicit submodule. We handle the sharing bit by doing a simple reference counting for the implicit submodules.

Reviewed By: dkgi

Differential Revision: D16031720

fbshipit-source-id: cffd60e6ed09da254fe1a30ad2f353421548165e