Clean up `Module` constructors
Summary:
We make the separation between non-implicit modules and implicit submodules a little bit more clear through constructors:
- `Module.create` creates a non-implicit module from a parsed `Source`.
- `Module.create_implicit` creates an implicit submodule.

Note that I cannot get rid of the old `Module.create` API since it is still relied upon for testing. In general I am always against this kind of abstraction-breaking API in tests as it usually indicates that the test code is poorly structured. But cleaning those bits up is too much of a deviation for me so I'll just rename it to `Module.create_for_testing` and call it for the day at the moment...

Reviewed By: dkgi

Differential Revision: D16032727

fbshipit-source-id: 45c1e96ffede21f456a1d943e344bb8ea7ba2b18