Handle inheriting from a class from a placeholder stub in a more principled manner
Summary:
The last blocker to making the type order only contain primitives is the appearance of Any.
The main source of these `Any`s are from inheriting from a class imported from a placeholder stub.
Because `Resolution.parse_annotation` "corrects" all types imported from a placeholder stub to be `Any`s, these classes are added as being subclasses of `Any` in the typeorder.
When I initially removed this functionality I got a bunch of errors in uses of `super` by these classes, and in uses of initializing these classes.
The problem was that by inheriting from `Any` and not nothing, these classes didn't have an edge to `object` added, but these two uses searched for parent classes, and just choked on `Any`.
These uses chose to fail silently, which I guess is fine, but now that we only have valid classes in the type order, they were throwing errors relating to trying to initialize `object` with too many parameters.

To solve this, I added a field to the `class_metadata` marking if this class inherited from a placeholder stub class, then if that were true, ignored the call to `super`, and also added a fake `__init__` method if one did not exist.

This metadata field can also be used to suppress missing attribute errors as has been desired for a while (T36048972).

Reviewed By: grievejia

Differential Revision: D16038914

fbshipit-source-id: b6da497c75c1be96e02c10d0af2e7bf13f5d30a0