Remove `File.Handle.t` in `TypeCheckResponse` message
Summary:
Not sure why we structure `TypeCheckResponse` as a map of [handle -> errors in the handle] in the first place. This mapping doesn't have to be eagerly constructed: we can simply have a list of errors there and construct the mapping from the list on-demand.

In fact, the only place where the map structure is necessary is when building one of the LSP response. The list structure seems to be in more popular demand in our codebase.

Reviewed By: sinancepel

Differential Revision: D16084685

fbshipit-source-id: 81eb216c160affc3c79a7b1642a3263d640f8ef1