Correctly handle expressions resolving to Any in types
Summary:
When something is annotated with an expression that resolves to Any, we ought not error, since Any is valid anywhere.
However, we should not accept that expression blindly as a primitive, as that causes analysis failures down the line.
Instead, we should suppress the error, but "correct" the type to Top.

Reviewed By: dkgi

Differential Revision: D16082111

fbshipit-source-id: 855a1748f25aee38f629e41b2b67b7934f2d8cfa