only generate models for top-level functions
Summary:
Previously, the view name logic would happily handle cases such as:

```
def foo():
  def bar(): ...
  return bar

urlpatterns = [url(callback=foo())]
```

by generating a model for `foo.bar`. This is not understandable to our static analysis, so we should disallow it.

As a side effect, the test functions had to be hoisted to the top level.

Reviewed By: dkgi

Differential Revision: D16091340

fbshipit-source-id: dddc19bb2e70d0c36955ab3cc4a5f1d5e6b66226