Add `__getattr__ : Callable[..., Any]` for all classes that inherit from a placeholder stub
Summary:
With the new metadata field introduced by D16082111, we can finally correctly suppress errors related to accessing properties inherited from placeholder stub classes.

I decided to tackle this with a `__getattr__` but would be open to changing it to be a direct error suppression.  I just thought that this would end up being more performant, and was more elegant.

Reviewed By: sinancepel

Differential Revision: D16093398

fbshipit-source-id: 71445392cb3b018a1dbdef398ef9b384207bcd84