Allow versioned_srcs to be an array
Summary:
Although not common, but it's possible that versioned sources is declared as an array like `["a.py", "b.py"]` which should be expanded into `{"a.py": "a.py", "b.py": "b.py"}`.
This commit fixes this problem so that when encountering an array instead of an object in versioned_srcs, buck builder will work and no longer crash.

Reviewed By: sinancepel

Differential Revision: D16111098

fbshipit-source-id: 6f6c01aa276886fce7a6f6d2d0a6514c722bf533