validate that the names of models are consistent for Pysa
Summary:
Right now, Pysa only validates that the arity of models for callables is consistent with the registered callables for Pyre.
We should make it so that names are also considered, and that we disallow models with inconsistent names for parameters.
This diff checks names of the parameters and also takes `*args` and `**kwargs` into consideration when a parameter name in the model is not found in the parameter name set of the type.
It is found that some of our existing tests do not declare the correct name. They are also fixed in this commit to make all tests pass.
It will also report a list of errors about where the inconsistencies are.

Reviewed By: fahndrich

Differential Revision: D16063154

fbshipit-source-id: 08aaf28ca54560afa9a7cdbd860735d394d3fceb