Partially back off D13541440
Summary:
Piggybacking on the location info is a really clever trick, but it also creates unnecessary dependencies on the `Handles` table, which is likely to go away in the future.

I wasn't able to come up with an easier way to eliminate the dependency on location info than re-scanning the source. We have to take a perf hit for this. Hopefully we'll bring it back when we refine the granularity of dependency analysis and eliminate the need for `~handle` in `Handler.register_alias`.

This diff is a pre-requsite of the diff above, which purges `File.Handle` from the environment.

Reviewed By: sinancepel

Differential Revision: D16132569

fbshipit-source-id: 9ec89c359b7ab9c2bcdb2fca046f2cbb6648682d