Adding first draft of subtyping rules
Summary:
This is a first draft at documenting the main component of the subtyping rule:
`less_or_equal`. We still need `join` and `meet` to have a realistic package.

The style is to write down the rules in the same order as the code. At some
point we might want to change to a style with only "positive rules" (e.g. don't
write anything about Top <= T, because if there is no rule to build it, we
can't prove it). That might evolve in the future

Reviewed By: dkgi

Differential Revision: D16120038

fbshipit-source-id: ddc1ee5df934771e1d3185bf98f3cae04e4f292d