document that class sinks and sources do not support merging with individual methods
Summary:
Class sinks and sources are meant to be used as one-liners, as in:

```
class MyClass(TaintSource[UserControlled]): ...
```

In the model tests for class models, we have some methods registered with taint sources to register them in the type environment.

Add some tests to clarify that these don't have any affect on class model generation.

Reviewed By: fahndrich

Differential Revision: D16094033

fbshipit-source-id: 13423cfadf0155d236bfd71d3abf6e4e2f508da6