use access paths for __setitem__
Summary:
Previously, when the assignment `x[a][b] = e` was encountered, we'd write store taint at the access path of `x`. This is imprecise, as it would lead to the following problematic case:

```
def foo():
  d = {}
  d[1] = __test_source()
  return d[2] # We mark this as a source of taint!
```

Store taint at a lower level, i.e., now keep track of the indices, and store taint at `d[1]` when encountering this case.

Reviewed By: fahndrich

Differential Revision: D16114299

fbshipit-source-id: ba94122c412a6cad96b5f57bbb0eef7f24633dc6