Separating the class hierarchy from the type order module
Summary:
Now that we have removed everything but classes from what was previously called the "type order" we can finally separate the functions that apply to that hierarchy from the functions that work on the complete type order, which is a lattice defined on all possible `Type.t`s, defined by both the class hierarchy and functions like `is_protocol` etc.

This makes the work ahead for expanding the set of generics to include variadics much simpler, since all of that work should be done on the class hierarchy, not the type order.

Overall, the strategy was that the functions that are now defined on `Type.Primitive` got moved into `ClassHierarchy`, and the ones that operated on `Type.t`s, remained in `TypeOrder`.

Those class hierarchy functions are like:
- connecting things
- adding things
- deduplication and disconnection
- lowest upper bound and greatest lower bound
- propagation of variables through the hierarchy

Those type order functions are the pure lattice operations
- meet/join
- leq
- solve leq

Reviewed By: sinancepel

Differential Revision: D16168700

fbshipit-source-id: 1bfe0514999c2c404da9905018df20934343a519