extend model checking to support unannotated attributes for classes
Summary:
Pyre is a bit finicky in how it deals with unannotated globals and attributes - it returns a annotation with `Type.Top` and global set to true. We were previously supporting global models by
doing an additional check of whether the name being validated was a global in the environment. This doesn't work for class attributes; instead, use the richer `Annotation.t` object that has information
on whether a type originates from a global.

This also reveals a bug in our attribute creation code - we probably don't want to mark non-existent attributes as immutable.

Reviewed By: fahndrich

Differential Revision: D16205525

fbshipit-source-id: ca3659c979e04c80dbb573ca45cd3fe7e2a2befe