Clean test setup, part IV
Summary:
Our test setup is very arbitrary at the moment.
- We create files at arbitrary location.
- We invoke `Service.Parser.parse_source_files` without consistently taking care of all the shared memory cleanups after we are done.
- We pay no attention to the consistency of `Configuration`. Mock configurations are created at-will even though some of their settings contradict with each other.
- We use `File.Handle.create` and `Reference.create` to attach arbitrary handles/qualifiers to files even though the filesystem setup is not properly matched up.
- In most cases we skip setting up `ModuleTracker` in the test, despite the fact that `ModuleTracker` is used in production now for end-to-end checks and is specifically designed to handle the previous bullet.

These issues are big pain points in test maintenance, especially when I try to expand the `ModuleTracker` integration.

This diff is a follow-up to D16207569, where we try to clean up `requestTest` and `autoCompleteTest`. The issue here is similar to `serverTest` we mentioned in D16213843, i.e. inconsistent mock states. The solution here is similar: use `CommandTest.ScratchServer.start` to make sure that every time we start a server, we start from scratch and deal with the real setup, not the mock one.

It turns out that connecting `ScratchServer` to all the tests ends up with almost a rewrite of `serverTest`!

Reviewed By: dkgi

Differential Revision: D16217084

fbshipit-source-id: 5c51e3e9b39eeffaea436c186cfb331cabd0d676